requirements.txt
python-telegram-bot==22.6
faster-whisper==1.2.1
ollama==0.6.1
Pillow==12.1.1
reportlab==4.4.10
python-dotenv==1.2.2
